Terminando trabalho

+++ Api-Semana1/academia/src/main/java/com/residencia/academia/controller/InstrutorController.java
@@ -16,35 +16,40 @@ import org.springframework.web.bind.annotation.RestController;
 
 import com.residencia.academia.dto.InstrutorDTO;
 import com.residencia.academia.entity.Instrutor;
+import com.residencia.academia.exception.NoSuchElementFoundException;
 import com.residencia.academia.service.InstrutorService;
 
 @RestController
 @RequestMapping("/instrutor")
 public class InstrutorController {
-	
+
 	@Autowired
 	private InstrutorService instrutorService;
 
 	@GetMapping
 	public ResponseEntity<List<Instrutor>> findAllInstrutor() {
 		List<Instrutor> instrutorList = instrutorService.listarTodos();
-		return new ResponseEntity<>(instrutorList, HttpStatus.OK);
-	}
-	
-	@GetMapping("/dto/{id}")
-	public ResponseEntity<InstrutorDTO> findInstrutorDTOById(@PathVariable Integer id) {
-		InstrutorDTO instrutorDTO = instrutorService.listarUmDTO(id);
-		return new ResponseEntity<>(instrutorDTO, HttpStatus.OK);
-
+		if (instrutorList.isEmpty()) {
+			throw new NoSuchElementFoundException("Lista vazia");
+		} else {
+			return new ResponseEntity<>(instrutorList, HttpStatus.OK);
+		}
 	}
 
 	@GetMapping("/{id}")
 	public ResponseEntity<Instrutor> findInstrutorById(@PathVariable Integer id) {
 		Instrutor instrutor = instrutorService.listarUm(id);
 		if (null == instrutor)
-			return new ResponseEntity<>(null, HttpStatus.NOT_FOUND);
+			throw new NoSuchElementFoundException("Não foi encontrado Instrutor com o ID: " + id);
 		else
 			return new ResponseEntity<>(instrutor, HttpStatus.OK);
+	}
+	
+	@GetMapping("/dto/{id}")
+	public ResponseEntity<InstrutorDTO> findInstrutorDTOById(@PathVariable Integer id) {
+		InstrutorDTO instrutorDTO = instrutorService.listarUmDTO(id);
+		
+		return new ResponseEntity<>(instrutorDTO, HttpStatus.OK);
 
 	}
 
@@ -52,7 +57,6 @@ public class InstrutorController {
 	public ResponseEntity<Instrutor> saveInstrutor(@RequestBody Instrutor instrutor) {
 		Instrutor novoInstrutor = instrutorService.saveInstrutor(instrutor);
 		return new ResponseEntity<>(novoInstrutor, HttpStatus.CREATED);
-
 	}
 
 	@PostMapping("/dto")
@@ -61,17 +65,29 @@ public class InstrutorController {
 		return new ResponseEntity<>(novoInstrutorDTO, HttpStatus.CREATED);
 
 	}
-	
+
 	@PutMapping
 	public ResponseEntity<Instrutor> updateInstrutor(@RequestBody Instrutor instrutor) {
 		Instrutor novoInstrutor = instrutorService.updateInstrutor(instrutor);
-		return new ResponseEntity<>(novoInstrutor, HttpStatus.OK);
+		if (null == novoInstrutor) {
+			return new ResponseEntity<>(null, HttpStatus.NOT_FOUND);
+		} else {
+			return new ResponseEntity<>(novoInstrutor, HttpStatus.OK);
+		}
 	}
 
 	@DeleteMapping("/{id}")
 	public ResponseEntity<String> deleteInstrutor(@PathVariable Integer id) {
-		instrutorService.deleteInstrutor(id);
-		return new ResponseEntity<>("", HttpStatus.OK);
+		Instrutor instrutor = instrutorService.listarUm(id);
+		if (null == instrutor) {
+			throw new NoSuchElementFoundException(
+					"Não foi possível excluir a Turma, pois não " + "foi encontrada uma turma com o ID: " + id);
+
+		} else {
+			instrutorService.deleteInstrutor(id);
+			return new ResponseEntity<>("Deletado com sucesso", HttpStatus.OK);
+		}
+
 	}
 
 }
